requirements.txt
requests
python-dateutil
google-genai